+++ service/src/test/java/ar/edu/itba/paw2019a2/service/RecipeServiceImplTest.java
@@ -40,7 +40,7 @@ public class RecipeServiceImplTest {
     private String INGREDIENT_NAME = "Ingredient";
     private boolean INGREDIENT_BOOLEANS = false;
     private double INGREDIENT_DOUBLES = 1.0;
-    private float INGREDIENT_FLOAT = (float)1.0;
+    private float INGREDIENT_FLOAT = 1.0f;
     private String INGREDIENT_TYPE_OF_SERVING = "Units";
     private int INGREDIENT_USER_ID = 1;
     private String INGREDIENT_STATUS = "REGULAR";
@@ -161,4 +161,6 @@ public class RecipeServiceImplTest {
         //2.
         recipeService.update(newRecipe);
     }
+
+    //getRecipesBasedOnOrderTagsCookable
 }
